Moving LockedObject template to its own header.
